requirements.txt
aiohttp==3.10.11
redis==5.0.8
web3==6.20.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
python-dotenv==1.0.1
openai==1.55.3
